requirements.txt
# Universal Quiz Solver API - Requirements

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests
requests==2.31.0

# Browser automation
playwright==1.40.0

# Environment variables
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# PDF processing
pdfplumber==0.10.3

# Image processing & OCR
Pillow==10.1.0
pytesseract==0.3.10

# Visualization
matplotlib==3.8.2

# Machine learning
scikit-learn==1.3.2
joblib==1.3.2

# HTML parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Progress bars
tqdm==4.66.1

# Async support
aiofiles==23.2.1

# Excel support
openpyxl==3.1.2
